Now, perturbation can have intermediate positions.

+++ MagnethoreologicalFluid2DWithWalls/analysis.h
@@ -9,7 +9,7 @@ private:
 	int particles_;
 	int length_;
 	int dimensions_;
-	int field_direction_;
+	double field_direction_;
 	int window_;
 	double pi = 3.14159265359;
 	double epsilon_ = 0.1;
@@ -36,7 +36,7 @@ private:
 	bool EndSimulation();
 
 public:
-	Analysis(double mason, double amplitude_relationship, int particles, int length, int window, int dimensions, int field_direction);
+	Analysis(double mason, double amplitude_relationship, int particles, int length, int window, int dimensions, double field_direction);
 	~Analysis();
 	bool PreAnalysis(double* x, double* y, double* z, double time);
 	void Connectivity(double* x, double* y, double* z);
