+++ MagnethoreologicalFluid2DWithWalls/dynamics.cpp
@@ -21,7 +21,9 @@
 
 using namespace std::chrono;
 using namespace cl;
-void Simulation(int field_direction, int phases, int particles, int dimensions, int length, double mason, double amplitude_relationship, double original_delta_t, int repetition, double max_times[3], bool keep_positions, bool load_positions, double creep_time) {
+void Simulation(double field_direction, int phases, int particles, int dimensions, int length, double mason, 
+	double amplitude_relationship, double original_delta_t, int repetition, double max_times[3],
+	bool keep_positions, bool load_positions, double creep_time) {
 	auto start = high_resolution_clock::now();
 	if (load_positions) {
 		std::cout << "Starting simulation for AR = " + std::to_string(amplitude_relationship) + ", Mason = " + std::to_string(mason) + ", field direction = " + std::to_string(field_direction) + ", creep time " + std::to_string(creep_time) + " and repetition " + std::to_string(repetition) + "\n";
@@ -150,7 +152,7 @@ void Simulation(int field_direction, int phases, int particles, int dimensions, 
 	cl::Buffer buffer_phase = cl::Buffer(context, CL_MEM_READ_ONLY, sizeof(int));
 	cl::Buffer buffer_dimensions = cl::Buffer(context, CL_MEM_READ_ONLY, sizeof(int));
 	cl::Buffer buffer_length = cl::Buffer(context, CL_MEM_READ_ONLY, sizeof(int));
-	cl::Buffer buffer_field_direction = cl::Buffer(context, CL_MEM_READ_ONLY, sizeof(int));
+	cl::Buffer buffer_field_direction = cl::Buffer(context, CL_MEM_READ_ONLY, sizeof(double));
 	cl::Buffer buffer_particles = cl::Buffer(context, CL_MEM_READ_ONLY, sizeof(int));
 	cl::Buffer buffer_delta_t = cl::Buffer(context, CL_MEM_READ_WRITE, sizeof(double));
 	cl::Buffer buffer_original_delta_t = cl::Buffer(context, CL_MEM_READ_ONLY, sizeof(double));
@@ -301,7 +303,7 @@ void Simulation(int field_direction, int phases, int particles, int dimensions, 
 	queue.enqueueWriteBuffer(buffer_particles, CL_TRUE, 0, sizeof(int), &particles);
 	queue.enqueueWriteBuffer(buffer_dimensions, CL_TRUE, 0, sizeof(int), &dimensions);
 	queue.enqueueWriteBuffer(buffer_length, CL_TRUE, 0, sizeof(int), &length);
-	queue.enqueueWriteBuffer(buffer_field_direction, CL_TRUE, 0, sizeof(int), &field_direction);
+	queue.enqueueWriteBuffer(buffer_field_direction, CL_TRUE, 0, sizeof(double), &field_direction);
 	queue.enqueueWriteBuffer(buffer_matrix_size, CL_TRUE, 0, sizeof(int), &matrix_size);
 	queue.enqueueWriteBuffer(buffer_delta_t, CL_TRUE, 0, sizeof(double), &delta_t);
 	queue.enqueueWriteBuffer(buffer_original_delta_t, CL_TRUE, 0, sizeof(double), &delta_t);
